Get custom hit info type to work
