+++ Reconstruction/RecDriftChamber/src/components/dictionary.xml
@@ -1,0 +1,6 @@
+<lcgdict>
+  <class name="hitINFO" >
+  </class>
+  <class name="std::vector<hitINFO>" >
+  </class>
+</lcgdict>
